fix(connection): split refresh function
